requirements.txt
streamlit>=1.36.0
requests>=2.32.0
python-dotenv>=1.0.1
pydantic>=2.7.0
